Segregando em fila a importação do acervo documental
