+++ SME.CDEP.Aplicacao/UseCase/Interface/IExecutarImportacaoArquivoAcervoDocumentalUseCase.cs
@@ -1,0 +1,8 @@
+
+using SME.CDEP.Infra;
+
+namespace SME.CDEP.Aplicacao
+{
+    public interface IExecutarImportacaoArquivoAcervoDocumentalUseCase : IRabbitUseCase
+    {}
+}
